mqtt-kafka-bridge/requirements.txt
paho_mqtt==2.1.0
pykafka==2.8.0
